Turn off `require-await` rule
I don't think we should use this rule - mainly for the reasons outlined
on the rule page itself; that using `async` is a useful shorthand for
ensuring a function returns a Promise in all scenarios, particularly
when an exception is thrown.

`async` also makes it very clear that a function returns a Promise,
when it might otherwise be unclear at a glance because of uncertainty
about what value is being `return`-ed.

There are some cases where we want to make a method `async` despite no
asynchronous work being performed, just so that it conforms to an
interface shared with other classes that _do_ need to be `async` (e.g.
the keyrings).