Add JSDoc ESLint rules
ESLint rules for JSDoc have been added to the base package and to the
TypeScript package, using the `eslint-plugin-jsdoc` plugin. These rules
should help improve the quality of our inline documentation.

This was done primarily to prepare for the generation of HTML
documentation using TypeDoc. Our inline documentation needs to be of a
higher quality than it is now for that to work out well.

These rules have been tested with `metamask-extension` [1],
`@metamask/controllers` [2], and `@metamask/eth-sig-util` [3]. See
those draft PRs to get an idea of what these rules look like in
practice.

[1]: https://github.com/MetaMask/metamask-extension/pull/12112
[2]: https://github.com/MetaMask/controllers/pull/591
[3]: https://github.com/MetaMask/eth-sig-util/pull/206