Fix the `lint` script (#204)
The `lint` npm script was not running the "miscellaneous" lints
correctly because the `--check` flag was missing. So instead of
checking for lint errors and returning non-zero if any problems were
found, it was printing all matched files to the console.

When the `--check` flag was restored, it became clear that the
`lint:fix` script could no longer pass because the `config-validation`
lint step was generating invalid snapshots. The `config-validation`
script was run last, and didn't run `prettier` on the rules before
writing them to disk.

`prettier` is now run on the snapshots before writing them to disk,
which ensures `lint:fix` and `lint` won't get into any more
disagreements.