Disable `@typescript-eslint/no-throw-literal` (#210)
The rule `@typescript-eslint/no-throw-literal` has been disabled
because it requires type information, which we hadn't realized when the
rule was added. We're saving rules that require type information for a
future release.

The rule that it extends (`no-throw-literal`) is left as disabled
because it doesn't work properly with TypeScript (hence the existence
of the extension rule).