Ignore rest siblings for `no-unused-vars`
The `ignoreRestSiblings` rule has been enabled for the `no-unused-vars`
rule [1] and the `@typescript-eslint/no-unused-vars` rule [2]. This
allows the use of unused variables alongside a rest operator. This
pattern can be useful for excluding variables from a group that you
don't want to use.

This option is already in use in the `@metamask/controllers`
repository [3].

[1]: https://eslint.org/docs/rules/no-unused-vars
[2]: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unused-vars.md
[3]: https://github.com/MetaMask/controllers/blob/8134da08a03fb63b146f8a3cffcb5e78fb70d58b/.eslintrc.js#L40