Remove `@typescript-config/consistent-type-definition` rule
The `@typescript-config/consistent-type-definition` rule has been
removed because it tries to force the use of interfaces over types in
scenarios where interfaces don't work due to incompatibilities with
`Record<string, unknown>` [1].

We could consider forcing the use of types over interfaces instead, but
I prefer using interfaces in some situations. They seem simpler to
write and understand sometimes. So for now it seems best to just remove
the rule.

This rule has already been disabled in the `@metamask/controllers`
repository [2].

[1]: https://github.com/Microsoft/TypeScript/issues/15300#issuecomment-702872440
[2]: https://github.com/MetaMask/controllers/blob/8134da08a03fb63b146f8a3cffcb5e78fb70d58b/.eslintrc.js#L36