Forbid TypeScript interfaces (#216)
TypeScript interfaces suck, they don't work with our `Json` type, and it's high time that we abolish them from every part of our codebase.

To wit, the only thing that can be done with an interface that cannot be done with a type is [declaration merging](https://www.typescriptlang.org/docs/handbook/declaration-merging.html), which is a "good language feature" in much the same way that a knife without a handle is a "good cooking instrument". All of the actually good stuff – IMO first and foremost `extends` and `implements` – [is also supported by object types](https://www.typescriptlang.org/docs/handbook/2/everyday-types.html#differences-between-type-aliases-and-interfaces) (in this case, via `&` and `implements`).

So, if we have two language features A and B, where the capabilities of A are a strict superset of those of B, and using B at all will sometimes cause uses supported by A to fail, the only sane thing to do is get rid of B. "B" is interfaces. Long live interfaces!