Upgrade all peer and dev dependencies
This commit was mainly motivated by the fact that in projects with
upgraded versions of TypeScript, TypeScript ESLint will produce a
warning when run. I decided to upgrade all dependencies while we at it
so that we can take advantage of any new features that the packages we
use may provide. This will, of course, compel us to upgrade those
packages in our projects, but considering that we don't need to upgrade
to a new version of `eslint-config` often, this can be done on a
case-by-case basis.

* Bump `@eslint/eslintrc` to ^1.3.0
* Bump `@lavamoat/allow-scripts` to ^2.0.3
* Bump `@metamask/auto-changelog` to ^2.6.0
* Bump `@typescript-eslint/eslint-plugin` and `@typescript-eslint/parser` to ^5.27.0
* Bump `eslint` to ^8.16.0
* Bump `eslint-config-prettier` to ^8.5.0
* Bump `eslint-plugin-import` to ^2.26.0
* Bump `eslint-plugin-jest` to ^26.4.6
* Bump `eslint-plugin-jsdoc` to ^38.1.6 (note: this is not the latest
  version, but 39.x drops support for Node 12)
* Bump `eslint-plugin-mocha` to ^9.0.0  (note: this is not the latest
  version, but 10.x drops support for Node 12)
* Bump `eslint-plugin-prettier` to ^4.0.0
* Bump `patch-package` to ^6.4.7
* Bump `prettier` to ^2.6.2
* Bump `typescript` to ^4.7.2

One note about ESLint is that we have to patch it in order to access the
config provided by the `eslint:recommended` preset, which is no longer
exposed in ESLint >= 8.x.