Update ESLint from v7 to v8
ESLint has been updated to v8 in all packages. This lets us use the
latest version of ESLint in our other packages without getting
`peerDependency` mismatch warnings.

The main change to be aware of is that four new recommended rules have
been added:
* `no-loss-of-precision` [1]
* `no-nonoctal-decimal-escape` [2]
* `no-unsafe-optional-chaining` [3]
* `no-useless-backreference` [4]

Our validation script was broken in this update because ESLint stopped
exporting their recommended configuration. `patch-package` was added so
that we could patch ESLint to export this configuration. This was taken
from #221, thanks @mcmire for this idea.

[1]: https://eslint.org/docs/latest/rules/no-loss-of-precision
[2]: https://eslint.org/docs/latest/rules/no-nonoctal-decimal-escape
[3]: https://eslint.org/docs/latest/rules/no-unsafe-optional-chaining
[4]: https://eslint.org/docs/latest/rules/no-useless-backreference