Allow async functions without any 'await' (#262)
The lint rules have been updated to allow functions to be `async` even
when they do not use `await`. It can be useful to make functions
`async` sometimes even when `await` is not used, because it can
simplify error handling for functions that return a Promise without
`await`-ing it (it ensures errors are always rejected, not thrown). It
can also be useful in cases when a function that does synchronous work
needs to be async to match an interface.