Stop requiring newlines between multiline blocks/expressions (#263)
This effectively reverts #197. This rule is being removed because we
have found that it forces us to introduce newlines in places where they
obstruct readability rather than helping it.

When grouping statements, it is often useful to "chunk" related lines
together by using a newline before and after the chunk, to denote that
each of the statements are related in some way. This is often done in
unit tests for example, when the "arrange", "act", "assert" grouping
style is used. This rule was forcing us to break up chunks.