Add rule `@typescript-eslint/prefer-enum-initializers` (#269)
The rule `@typescript-eslint/prefer-enum-initializers` has been added.
This rule ensures that we use initializers for enums.

Uninitialized enums are hazardous because they implicitly assign each
entry a sequential number value. This means that adding a new entry to
the enum changes the value of every following entry. This can be
confusing and unexpected, especially if this enum is stored in
persisted state and now has a different meaning than before.
Initialized enums, on the other hand, never unexpectedly change values.

This is a breaking change.