Disable @typescript-eslint/no-unnecessary-type-assertion
This rule, as we've discovered after running it against the `core`
monorepo, seems to produce false negatives. If you look at the issues
list for the `typescript-eslint` repo and filter by this rule, you'll
notice that [other people have reported this as well][1]. As a result,
it's not worth enabling right now.

[1]: https://github.com/typescript-eslint/typescript-eslint/issues?q=is%3Aissue+is%3Aopen+no-unnecessary-type-assertion+sort%3Aupdated-desc