Bump min Node to 16, and use LTS Node for dev
**This change is breaking.**

We could bump this monorepo to Node 18, but since these packages are
heavily depended upon, we don't need to do that yet until we've
confirmed all other projects are using 18.

Using the LTS version of Node puts us on a more modern version for
development (and also happens to fix the `.nvmrc` file so `asdf` can
read it).