Disable TS no-unsafe-enum-comparison rule
The authors of the `@typescript-eslint/no-unsafe-enum-comparison` rule
seem to think that enums should always be treated as a set of opaque
labels rather than an object container of known values or mere shortcuts
for string values. This means that the following is disallowed:

``` typescript
// The lint error occurs here (note `namespace` is a `string`)
switch (namespace) {
  case KnownCaipNamespace.Wallet:
    if (reference === KnownCaipNamespace.Eip155) {
      return isSupportedEip155Account();
    }
    return isSupportedNonEvmAccount();
  case KnownCaipNamespace.Eip155:
    return isSupportedEip155Account();
  default:
    return isSupportedNonEvmAccount();
}
```

However, this pattern is very common across our code, so at the moment
we get all kinds of errors.

While there is [past discussion on this issue][1] which explains that
this is an unsafe operation because there's no guarantee that the string
has any overlap with the enum, **whether or not the string overlaps is
the very thing we are trying to test**. And while it is true that enum
values could technically be anything, we are making what I think is a
reasonable assumption that enum values are well-named, so
`KnownCaipNamespace.Eip155` equates to something like "eip155" and not
"asdfadfdafasdlf".

Since we're slowly moving away from enums, I think we should disable
this rule.

[1]: https://github.com/typescript-eslint/typescript-eslint/issues/8557