feat(typescript): Update type import specifier rules (#381)
Ref: https://github.com/MetaMask/ocap-kernel/pull/175

This replaces `@typescript-eslint/consistent-type-imports` with `import-x/consistent-type-specifier-style`, and an exhortation to enable `verbatimModuleSyntax` (requires TypeScript >=5) in the tsconfigs of our packages.

Heretofore, we have enforced the existence of type import specifiers, but not that they are stylistically consistent. The options are:
- "top-level", e.g. `import type { a } from 'x'}`
- "inline", e.g. `import { type a } from 'x'}`
 
The rule `@typescript-eslint/consistent-type-imports` has been responsible for this. This rule has an option named `fixTypes`, but that only applies when a type-only import is _not_ specified as a type-only import. In other words, it enforces the existence of type import specifiers, but not that they're stylistically consistent.

Rather than using `@typescript-eslint/consistent-type-imports` and its confusingly named options, we are better off enabling `verbatimModuleSyntax` in our `tsconfig` files, and enabling `import-x/consistent-type-specifier-style`. In this way, the existence of type import specifiers is enforced by `tsc`, and the `import-x` ESLint plugins ensures that they are consistent. We use the "top-level" style because that results in `import type { a } from 'x';` being elided completely from the JS output as opposed to `import { type a } from 'x';` which would be emitted as `import {} from 'x';`. I don't know if that's actually harmful, but it seems surprising, and I don't like surprises.

Supporting documentation:
- `@typescript-eslint/consistent-type-imports` [rule documentation](https://typescript-eslint.io/rules/consistent-type-imports/#comparison-with-importsnotusedasvalues--verbatimmodulesyntax)
- [The `verbatimModuleSyntax` docs](https://www.typescriptlang.org/tsconfig/#verbatimModuleSyntax)
- This tseslint issue: https://github.com/typescript-eslint/typescript-eslint/issues/9652