Support TypeScript 5.6.x-5.8.x (#395)
This commit loosens the peer restriction on `typescript` to allow
projects to upgrade to more recent versions of TypeScript (such as 5.8,
the latest version at the time of writing). Similarly, the peer
restriction on `typescript-eslint` has been loosened to support more
recent versions.

Both of these packages have been updated in development as well.