Disable `@typescript-eslint/no-unsafe-enum-comparison` rule (#387)
This disables the `@typescript-eslint/no-unsafe-enum-comparison` rule,
which disallows comparing enums to string or number values. For example,
the following is not allowed when this rule is enabled:

```ts
enum Foo {
  Bar = 'Bar',
  Baz = 'Baz',
}

const myVariable: string = "Bar";
if (myVariable == Foo.Bar) {
  // ...
}
```

We use this pattern in multiple places, so I think this rule is too
strict and we should disable it.