Widen allowed TS / TS ESLint versions
The latest version of TypeScript is 5.9, but no project can currently
use this without producing peer dependency warnings. To prevent this and
unlock new versions of TypeScript in the future, widen the peer
dependencies:

- Allow TypeScript 5.9 and any future version up to 6
- Allow TypeScript ESLint versions betwen 8.24 and 9

To ensure this does not cause a problem in this repo, also update the
development version of TypeScript and TypeScript ESLint to their latest
versions.