Allow remaining TypeScript 5.x versions (#422)
The latest version of TypeScript is 5.9, but no project can currently
use this without producing peer dependency warnings. To prevent this and
unlock new versions of TypeScript 5.x in the future, widen the peer
dependency.

To ensure this does not cause a problem in this repo, also update the
development version of TypeScript and TypeScript ESLint to their latest
versions.