chore: Increase minimum `typescript-eslint` version
The minimum version of `typescript-eslint` has been updated to v8.39.0.
This is the version that included widening the range of allowed
`typescript` versions, so this new minimum will ensure users of this
config won't see warnings about incompatible typescript versions if
they satisfy this new minimum.

This update also brings in a bug fix we encountered on `core` related
to the `@typescript-eslint/no-unnecessary-type-assertion` rule, which
was in v8.31.0. Details here:
https://github.com/typescript-eslint/typescript-eslint/issues/8721