chore: Increase minimum `typescript-eslint` version (#423)
The minimum version of `typescript-eslint` has been updated to v8.39.0.
This is the version that included widening the range of allowed
`typescript` versions, so this new minimum will ensure users of this
config won't see warnings about incompatible typescript versions if they
satisfy this new minimum.

This update also brings in a bug fix we encountered on `core` related to
the `@typescript-eslint/no-unnecessary-type-assertion` rule, which was
in v8.31.0. Details here:
https://github.com/typescript-eslint/typescript-eslint/issues/8721

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> Increase `typescript-eslint` peer dependency minimum to ^8.39.0 and
document as a breaking change.
> 
> - **TypeScript config package (`packages/typescript`)**:
> - **Peer deps**: Raise `typescript-eslint` minimum from `^8.24` to
`^8.39.0` in `package.json`.
> - **Changelog**: Add Unreleased entry marking this as a **BREAKING**
change with notes on compatibility and fixes.
> 
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
735688e5c57cdf6796873245627c329a0a4addc7. This will update automatically
on new commits. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->