feat: Convert all warnings to errors
We try to avoid using ESLint warnings because they tend to be easy for
conitrbutors to miss, and they pile up over time and make the linter
miserable to work with. Our team's preference is to ignore or suppress
violations that we don't want to resolve directly.

We never set any rules to `warn` directly, but some of the configs
ew're inheriting from include warnings. All of them have been converted
to errors.