fix: Disable `jsdoc/check-indentation` to fix bullet list formatting
Bulleted lists and other types of indented sections are incorrectly
flagged by the current version of the `jsdoc/check-indentation` rule,
so it has been temporarily disabled.

We can bring it back in a later version after updating the plugin to
61.4, which includes a new option that can fix this problem. That has
been postponed for now because it requires bumping the minimum
supported Node.js version.

Fixes #404